refactor(sedona-gdal): remove wrapper re-export aliases
