+++ c/sedona-gdal/src/raster.rs
@@ -17,5 +17,3 @@
 
 pub mod rasterband;
 pub mod types;
-
-pub use rasterband::{actual_block_size, RasterBand};
